+++ projects/ngx-ukis-ui-clarity/schematics/add-routing/schema.ts
@@ -1,0 +1,6 @@
+export interface UkisNgAddRoutingSchema {
+  // The name of the angular project you want to add @dlr-eoc/ngx-ukis-ui-clarity.
+  project?: string;
+  addFiles?: boolean;
+  updateFiles?: boolean;
+}