+++ projects/ngx-ukis-ui-clarity/src/lib/global-progress/global-progress.component.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'ngx-global-progress',
+  imports: [],
+  templateUrl: './global-progress.component.html',
+  styleUrl: './global-progress.component.scss'
+})
+export class GlobalProgressComponent {
+
+}